+++ src/main/java/com/example/yamyam16/menu/dto/MenuListResponseDto.java
@@ -11,6 +11,6 @@ public class MenuListResponseDto {
 
 	private String menuName;
 
-	private int menuPrice;
+	private Long menuPrice;
 
 }
