+++ src/main/java/com/example/yamyam16/menu/dto/MenuUpdateRequestDto.java
@@ -9,5 +9,5 @@ public class MenuUpdateRequestDto {
 
 	private String menuName;
 
-	private int menuPrice;
+	private Long menuPrice;
 }
